added legal option to the menu and made VarantSpecificUtils an implementation of an Abstract Utils class
