+++ app/src/main/java/org/eyeseetea/malariacare/layout/adapters/survey/CompositeScoreAdapter.java
@@ -21,10 +21,8 @@ package org.eyeseetea.malariacare.layout.adapters.survey;
 
 import android.app.Activity;
 import android.content.Context;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.BaseAdapter;
 import android.widget.ListView;
 
 import org.eyeseetea.malariacare.R;
@@ -32,10 +30,8 @@ import org.eyeseetea.malariacare.database.model.CompositeScore;
 import org.eyeseetea.malariacare.database.model.Tab;
 import org.eyeseetea.malariacare.layout.score.ScoreRegister;
 import org.eyeseetea.malariacare.layout.utils.LayoutUtils;
-import org.eyeseetea.malariacare.utils.Utils;
+import org.eyeseetea.malariacare.utils.AUtils;
 import org.eyeseetea.malariacare.views.CustomTextView;
-
-import java.util.List;
 
 /**
  * Created by Jose on 21/04/2015.
@@ -84,7 +80,7 @@ public class CompositeScoreAdapter extends ATabAdapter {
         if (compositeScoreValue == null)
             ((CustomTextView)rowView.findViewById(R.id.score)).setText(getContext().getString(R.string.number_zero));
         else
-            ((CustomTextView)rowView.findViewById(R.id.score)).setText(Utils.round(compositeScoreValue));
+            ((CustomTextView)rowView.findViewById(R.id.score)).setText(AUtils.round(compositeScoreValue));
 
         rowView.setBackgroundResource(LayoutUtils.calculateBackgrounds(position));
