+++ app/src/main/res/menu/menu_dashboard.xml
@@ -18,14 +18,16 @@
   -->
 
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
-  <item android:id="@+id/action_settings" android:title="@string/settings_menu_configuration"
-      android:orderInCategory="100"/>
-  <item android:id="@+id/action_license" android:title="@string/settings_menu_licence"
-      android:orderInCategory="102"/>
-  <item android:id="@+id/action_about" android:title="@string/settings_menu_about"
-      android:orderInCategory="103"/>
-  <item android:id="@+id/action_pull" android:title="@string/settings_menu_pull"
-      android:orderInCategory="103"/>
-  <item android:id="@+id/action_logout" android:title="@string/settings_menu_logout"
-      android:orderInCategory="105"/>
+    <item android:id="@+id/action_settings" android:title="@string/settings_menu_configuration"
+        android:orderInCategory="100"/>
+    <item android:id="@+id/action_license" android:title="@string/settings_menu_licence"
+        android:orderInCategory="102"/>
+    <item android:id="@+id/action_about" android:title="@string/settings_menu_about"
+        android:orderInCategory="103"/>
+    <item android:id="@+id/action_legal" android:title="@string/settings_menu_legal"
+        android:orderInCategory="104"/>
+    <item android:id="@+id/action_pull" android:title="@string/settings_menu_pull"
+        android:orderInCategory="105"/>
+    <item android:id="@+id/action_logout" android:title="@string/settings_menu_logout"
+        android:orderInCategory="106"/>
 </menu>
