+++ app/src/main/res/menu/menu_general.xml
@@ -24,6 +24,8 @@
         android:orderInCategory="102"/>
     <item android:id="@+id/action_about" android:title="@string/settings_menu_about"
         android:orderInCategory="103"/>
-    <item android:id="@+id/action_logout" android:title="@string/settings_menu_logout"
+    <item android:id="@+id/action_legal" android:title="@string/settings_menu_legal"
         android:orderInCategory="104"/>
+    <item android:id="@+id/action_logout" android:title="@string/settings_menu_logout"
+        android:orderInCategory="105"/>
 </menu>
