+++ app/src/main/res/values-pt/strings.xml
@@ -16,6 +16,7 @@
 
     <!-- settings -->
     <string name="settings_menu_about">Acerca de</string>
+    <string name="settings_menu_legal">Legal</string>
     <string name="settings_menu_configuration">Definições</string>
     <string name="settings_menu_licence">Licença</string>
     <string name="settings_menu_update">Actualizar questões (SISD) Sistemas de Informação de Saúde Distrital</string>
