+++ app/src/main/res/values/strings.xml
@@ -23,6 +23,7 @@
 
     <!-- settings -->
     <string name="settings_menu_about">About</string>
+    <string name="settings_menu_legal">Legal</string>
     <string name="settings_menu_configuration">Settings</string>
     <string name="settings_menu_licence">License</string>
     <string name="settings_menu_update">Update questions (DHIS)</string>
